Faz 2 (1/n): domain katmanı KMP ortak modülüne taşındı (#15)
* Faz 2 (1/n): domain katmanı KMP ortak modülüne taşındı

Android uygulaması ve ileride yazılacak iOS uygulaması artık **aynı** iş
kurallarını kullanıyor: para aritmetiği, fiyatlandırma, üyelik durumu, hakediş
oranı, seans kotası ve telefon doğrulaması `:shared` modülünde bir kez yazılı.
İki platformda ayrı yazılsalardı kaçınılmaz olarak saparlardı — bu projede
düzeltilen hataların önemli bir kısmı (kesir/yüzde karışması, `-1` sentinel'i,
serbest metin durum değerleri) tam olarak bu tür sapmalardı.

Domain katmanının platforma bağlı yalnızca üç noktası vardı ve üçü de çözüldü:

- `java.util.UUID` -> `expect`/`actual` `randomUuid()`. Android'de `UUID`,
  iOS'ta `NSUUID`. `NSUUID` büyük harf ürettiği için küçük harfe çevriliyor:
  kimlikler iki platform arasında senkronize olacak, biçim birebir aynı olmalı.
- `Math.round` -> `roundToLong()` / `roundToInt()`. İkisi de yarıyı yukarı
  yuvarlıyor, davranış değişmedi.
- `java.time` -> `kotlinx-datetime`. `Periods` API'si artık **epoch millis**
  konuşuyor; her platform kendi tarih tipini yalnızca kendi UI sınırında
  kullanıyor. `java.time` Android ekranlarında kaldı, domaine sızmıyor.

Testler `commonTest`e taşındı ve `kotlin.test`e çevrildi. Aynı testler hem JVM
hem Kotlin/Native üzerinde koşuyor: JVM'de geçip Native'de patlayan bir davranış
farkı (yuvarlama, tarih, metin) artık CI'da yakalanır.

CI'a macOS işi eklendi: `linkDebugFrameworkIosSimulatorArm64`,
`linkDebugFrameworkIosArm64` ve `iosSimulatorArm64Test`. Kotlin/Native derlemesi
yalnızca macOS'ta yapılabildiği için, "iOS'ta çalışıyor" iddiasını doğrulayan
tek yer burası; bu iş olmadan iddia doğrulanmamış kalırdı.

Ayrıca iki ölü dosya kalktı:
- `AppointmentStatus` — Faz 1'de `AppointmentState` enum'u yerini almıştı.
- `RateTest` — bir önceki dilimde eklerken `MoneyTest`in aynı davranışları zaten
  test ettiğini gözden kaçırmışım; yalnızca gerçekten yeni olan iki durum
  (sıfır matrah, küsuratlı oran) `MoneyTest`e taşındı.


* KMP için araç zincirini yükselt ve ölü WorkManager kurulumunu kaldır

CI kırmızıydı: Gradle 9.3.1 üzerinde Kotlin 2.0.21'in multiplatform eklentisi,
Gradle 9'da kaldırılmış bir iç sınıfı (`DefaultArtifactPublicationSet`) arıyor ve
`:shared` modülü konfigürasyon aşamasında düşüyordu. Android eklentisi bu sınıfa
dokunmadığı için uygulama modülü 14 PR boyunca sorunsuz derlenmişti; sorun
yalnızca KMP yolunda ortaya çıktı.

Kotlin 2.0.21 -> 2.2.21, KSP 2.0.21-1.0.28 -> 2.2.21-2.0.5, Hilt 2.51.1 -> 2.60.1.
Tavanı KSP belirledi: Kotlin 2.3 ve 2.4 için yayınlanmış bir KSP sürümü yok,
dolayısıyla Room ve Hilt kod üretimini kullanan bu proje en fazla 2.2.21'e
çıkabiliyor.

Yükseltmeyi ayrı bir PR'a bölmedim çünkü tek başına bir amacı yok — KMP için
gerekli olduğundan aynı değişikliğin parçası. İki farklı hata sınıfı da loglarda
birbirinden ayırt edilebiliyor (konfigürasyon hatası vs. kod üretimi hatası).

Bununla birlikte bir önceki dilimde ölü kalan kurulum da kalktı:
`MemberExpirationWorker` silindiğinde geriye hiçbir işçisi olmayan bir
`HiltWorkerFactory`, manifest'te kapatılmış bir auto-init ve üç kullanılmayan
bağımlılık (`hilt-work`, androidx `hilt-compiler`, `work-runtime-ktx`) kalmıştı.
Bunlar aynı zamanda yükseltmenin en riskli uyumluluk yüzeyiydi; kaldırılmaları
hem ölü kodu temizliyor hem de riski düşürüyor. Faz 4'teki senkronizasyon
kuyruğu gerçekten bir işçiye ihtiyaç duyduğunda kurulum onu kullanan kodla
birlikte geri gelecek.

Ayrıca kullanımdan kalkan `android { kotlinOptions { } }` bloğu Kotlin
eklentisinin `compilerOptions` alanına taşındı.


* Hilt sürümünü AGP 8.13 ile uyumlu olana çek

2.60.1 fazla ileriydi: "The Hilt Android Gradle plugin is only compatible with
Android Gradle plugin (AGP) version 9.0.0 or higher (found 8.13.2)" diyerek
eklenti uygulanırken düştü ve her iki CI işi de konfigürasyon aşamasında bitti.

Sürümü tahmin etmek yerine eklenti jar'larındaki minimum AGP kontrolünü okudum:

  2.53.1 / 2.55 / 2.56.2 -> AGP 8.1.0+
  2.57.2 / 2.58          -> AGP 8.4.0+
  2.59.2                 -> AGP 9.0.0+

2.58, elimizdeki AGP 8.13.2 ile çalışan en yeni sürüm; KSP2 desteği için de
fazlasıyla yeterli. AGP'yi 9'a çıkarmak alternatifti ama bu, ana sürüm atlayan
ve KMP geçişiyle hiç ilgisi olmayan ayrı bir iş olurdu.


* Test iddialarında argüman sırasını kotlin.test kuralına göre düzelt

JUnit ile kotlin.test arasında mesaj argümanının yeri farklı: JUnit'te **ilk**
(`assertEquals(message, expected, actual)`), kotlin.test'te **son**
(`assertEquals(expected, actual, message)`). commonTest'e taşırken import'ları
toplu değiştirmiştim; bu iki çağrının anlamı sessizce değişti:

  assertEquals("Girdi: 5551112233", "+905551112233", ...)
      -> mesajı beklenen değer sanıp karşılaştırdı
  assertNull("Girdi: ...", ...)
      -> mesaj metninin null olmasını bekledi, yani her zaman düştü

Kod tarafında bir sorun yok; `PhoneNumber` davranışı değişmedi. Bu, taşımanın
kendi getirdiği bir hataydı ve testler tam da bunun için var — sessizce
geçmeyip yakaladılar.

Diğer test dosyalarını da taradım: mesaj alan başka çağrı yok.

Bu turda araç zinciri tarafı ilk kez doğrulandı: `:shared:compileDebugKotlinAndroid`
geçti, yani Kotlin 2.2.21 + KMP eklentisi Gradle 9.3.1 üzerinde çalışıyor ve
65 testin 63'ü geçiyor.


* Ortak kodda JvmInline için açık import ekle

Kotlin/Native derlemesi düştü:

  e: Money.kt:16:2 Unresolved reference 'JvmInline'
  e: Money.kt:106:2 Unresolved reference 'JvmInline'

`kotlin.jvm.*` yalnızca JVM hedeflerinde varsayılan import. Kod app modülündeyken
(Android/JVM) `@JvmInline` otomatik çözülüyordu; ortak koda taşınınca açık import
gerekti. Android tarafı bu eksikliği hiç göstermedi.

Bu, macOS CI işini eklemenin somut karşılığı: iddiayı doğrulamak için koyduğum
iş, ilk çalıştığında gerçek bir taşınabilirlik hatası buldu. Yalnızca JVM'de
derleseydik `:shared` "ortak" görünmeye devam edecek, hata iOS uygulaması
yazılırken ortaya çıkacaktı.

Ortak kaynakları diğer JVM'e özgü ek açıklamalar için de taradım; `@JvmInline`
dışında kullanım yok.

Aynı turda Linux işindeki ortak modül testleri geçti: argüman sırası düzeltmesi
sonrası 65 testin tamamı yeşil.


---------

+++ app/src/main/java/com/gymapp/presentation/finance/FinanceViewModel.kt
@@ -72,7 +72,7 @@ class FinanceViewModel @Inject constructor(
         val nowMs = System.currentTimeMillis()
 
         /** Son [monthsBack] aylık tahsilat. Üst sınır `nowMs`: ileri tarihli kayıt ciroyu şişiremez. */
-        fun revenue(monthsBack: Long): Money {
+        fun revenue(monthsBack: Int): Money {
             val window = Periods.lastMonths(monthsBack, nowMs)
             return entries
                 .filter { it.isIncome && !it.isPending && !it.isVoided && it.occurredAtMs in window }
@@ -105,10 +105,10 @@ class FinanceViewModel @Inject constructor(
             totalIncome = income,
             totalExpense = expense,
             totalProfit = income - expense,
-            monthlyRevenue = revenue(1L),
-            quarterlyRevenue = revenue(3L),
-            halfYearlyRevenue = revenue(6L),
-            yearlyRevenue = revenue(12L),
+            monthlyRevenue = revenue(1),
+            quarterlyRevenue = revenue(3),
+            halfYearlyRevenue = revenue(6),
+            yearlyRevenue = revenue(12),
             selectedMonth = month,
             selectedYear = year,
             selectedFilter = filter,
